client: app actions, go cmd built
command processing now returns a vec of AppAction, allows mutation of
the app to be performed after all processing

the Go command builds out a model::Action to be sent to the frontend

next: create a channel between client::net and client::app to allow
message passing to the server. limited to model::Action for now, maybe.